pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-auto-skill"
version = "3.0.0"
description = "Automatically learn from your workflows and turn them into intelligent, context-aware skills"
readme = "README.md"
authors = [
    {name = "MaTriXy"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["claude", "ai", "skills", "automation", "patterns", "claude-code"]
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
v2 = [
    "pygls>=1.3.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
all = [
    "claude-auto-skill[v2,dev]",
]

[project.urls]
Homepage = "https://github.com/MaTriXy/claude-auto-skill"
Documentation = "https://MaTriXy.github.io/claude-auto-skill/"
Repository = "https://github.com/MaTriXy/claude-auto-skill"
Issues = "https://github.com/MaTriXy/claude-auto-skill/issues"
Changelog = "https://github.com/MaTriXy/claude-auto-skill/blob/main/CHANGELOG.md"

[project.scripts]
auto-skill = "commands.cli:main"
# Legacy entry points (deprecated, use `auto-skill <subcommand>` instead)
auto-skill-init = "commands.init:main"
auto-skill-discover = "commands.discover:main"

[tool.setuptools]
packages = ["core", "core.providers", "commands", "hooks", "scripts"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.json", "*.yaml"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=core",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_mental: requires Mental CLI to be installed",
    "requires_network: requires internet connection",
]
